Moved all test files to the __test__ directory. Removed common web calls to separate file webConn.js and updated all tests to use this common file.
