+++ __test__/UC10.test.js
@@ -1,5 +1,8 @@
+jest.dontMock('./webConn');
+const webConn = require('./webConn');
+
 beforeAll((done) => {
-    getFromWeb('http://localhost/puzzles.html', function(inString) {
+    webConn.getFromWeb('http://localhost/puzzles.html', function(inString) {
         document.body.innerHTML = inString;
         done();
     });
@@ -49,7 +52,7 @@ describe('Testing UC10 UI, displaying a puzzle for the user', () => {
         }
         answerButtonElem.click();
 
-        getFromApi('http://localhost/puzzlesAPI.php', requestObj, function(jsonObj) {
+        webConn.getFromApi('http://localhost/puzzlesAPI.php', requestObj, function(jsonObj) {
             expect(answerInput).not.toBe("");
             expect(requestObj.params.answer).not.toBe(""); // idk about this one but trying it out
             done();
@@ -70,7 +73,7 @@ describe('Testing UC10 API, displaying a puzzle for the user', () => {
             "params": "puzzle_id"
         }
 
-        getFromApi('http://localhost/', requestObj, function(jsonObj) {
+        webConn.getFromApi('http://localhost/', requestObj, function(jsonObj) {
             expect(jsonObj.dataset).not.toBe("");
             done();
         });
@@ -81,7 +84,7 @@ describe('Testing UC10 API, displaying a puzzle for the user', () => {
             "route": "get_puzzle",
             "params": "puzzle_id", 
         }
-        getFromApi('http://localhost/', requestObj, function(jsonObj) {
+        webConn.getFromApi('http://localhost/', requestObj, function(jsonObj) {
             expect(jsonObj.puzzle_id).not.toBe("");
             done();
         });
@@ -97,56 +100,9 @@ describe('Testing UC10 API, displaying a puzzle for the user', () => {
         }
         console.log(requestObj);
         answerButtonElem.click();
-        getFromApi('http://localhost/', requestObj, function(jsonObj) {
+        webConn.getFromApi('http://localhost/', requestObj, function(jsonObj) {
             expect(jsonObj.answer).not.toBe("");
             done();
         });
     });
 });
-
-
-jest.dontMock('http');
-const http = require('http');
-function getFromWeb(url, requestObj, callMeBack) {
-    http.post(url, function(response) {
-        let buffer = '';
-        response.on('data', function(piece) {
-            buffer += piece;
-        });
-        response.on('end', function() {
-            const jsonObj = JSON.parse(buffer);
-            callMeBack(jsonObj);
-        });
-    })
-};
-
-
-function getFromApi(method, url, requestString, callMeBack) {
-    if (method === "GET") {
-        getFromWeb(url, function(inString) {
-            var jsonObj = JSON.parse(inString);
-            callMeBack(jsonObj);
-        });
-    }
-    else {
-        var options = {
-            hostname: 'localhost',
-            port: 80,
-            path: url,
-            method: 'POST',
-            headers: { 'Content-Type': 'application/json', }
-        };
-        var request = http.request(url, function(response) {
-            let buffer = '';
-            response.setEncoding('utf8');
-            response.on('data', function(piece) {
-                buffer += piece;
-            });
-            response.on('end', function() {
-                callMeBack(buffer);
-            });
-        });
-        request.write(requestString);
-        request.end();
-    }
-};
