+++ __test__/UC9.test.js
@@ -1,5 +1,8 @@
+jest.dontMock('./webConn');
+const webConn = require('./webConn');
+
 beforeAll((done) => {
-    getFromWeb('http://localhost/puzzles.html', function(inString) {
+    webConn.getFromWeb('http://localhost/puzzles.html', function(inString) {
         document.body.innerHTML = inString;
         done();
     });
@@ -29,7 +32,7 @@ describe('Testing UC9 UI, Editing profile', () => {
         require('../editprofile');
         confirmButtonElem.click();
 
-        getFromApi('http://localhost/usersAPI.php', function(jsonObj) {
+        webConn.getFromApi('http://localhost/usersAPI.php', function(jsonObj) {
             expect(jsonObj).toEqual({
                 "username": "danni",
                 "email": "danni@danni.com",
@@ -57,7 +60,7 @@ describe('Testing UC9 API, Editing profile', () => {
             }
         };
 
-        getFromApi('http://localhost/usersAPI.php', requestObj, function(jsonObj) {
+        webConn.getFromApi('http://localhost/usersAPI.php', requestObj, function(jsonObj) {
             console.log(jsonObj);
             expect(jsonObj).toEqual({
                 "username": "danni",
@@ -70,49 +73,3 @@ describe('Testing UC9 API, Editing profile', () => {
         });
     });
 });
-
-jest.dontMock('http');
-const http = require('http');
-function getFromWeb(url, requestObj, callMeBack) {
-    http.post(url, function(response) {
-        let buffer = '';
-        response.on('data', function(piece) {
-            buffer += piece;
-        });
-        response.on('end', function() {
-            const jsonObj = JSON.parse(buffer);
-            callMeBack(jsonObj);
-        });
-    })
-};
-
-
-function getFromApi(method, url, requestString, callMeBack) {
-    if (method === "GET") {
-        getFromWeb(url, function(inString) {
-            var jsonObj = JSON.parse(inString);
-            callMeBack(jsonObj);
-        });
-    }
-    else {
-        var options = {
-            hostname: 'localhost',
-            port: 80,
-            path: url,
-            method: 'POST',
-            headers: { 'Content-Type': 'application/json', }
-        };
-        var request = http.request(url, function(response) {
-            let buffer = '';
-            response.setEncoding('utf8');
-            response.on('data', function(piece) {
-                buffer += piece;
-            });
-            response.on('end', function() {
-                callMeBack(buffer);
-            });
-        });
-        request.write(requestString);
-        request.end();
-    }
-};
